Adding limited Paramesh-like HDF5 binary IO capability
This commit adds the capability to write simulation data to HDF5 files
from a simulation compiled with Amrex (in "Mode 3"), in a form that is
somewhat compatible with HDF5 files created by PARAMESH-based FLASH
(traditional FLASH or "Mode 1" or "Mode 2").

This IO capability is "limited" in that some metainformation data that
should be present in a valid FLASH plotfile or checkpoint file will be
missing or zero-filled.  In particular, so-called checkpoint files
written with this capability cannot really be used for restarting a
simulation.

In the current implementation, only serial I/O is supported.
The current implementation only supports cell-centered ("UNK") data.
(It may not even compile if a simulation has face variables.)

In the current implementation, checkpoint files are written with
the new capability. Plot files are still written in the previous
way derived from AMReX I/O facilities. The new AM/io_writeData
implementation should be able to handle plotfiles, too, but this
is not tested.

To used, the IO/IOMain/hdf5/serial/AM directory should be explicitly
included in a configuration. (Perhaps create a shortcut.) Make sure
no conflicting PM or UG or parallel implementation of IO is requested.

Tested successfully with 2D Sedov + unsplit Hydro in Mode 3.

The resulting files work with sfocu.

Brief attempts to visualize such files with FIDLR and VisIt also had
at least partial success; VisIt showed pseudocolor plot mostly
normally but with some artifacts.