Use MPI_Get_Address instead of MPI_Address unless FLASH_MPI1 is defined
Also use MPI_TYPE_CREATE_STRUCT instead of MPI_TYPE_STRUCT.

This is a fix for use of old (MPI-1) interfaces. It prevents some
gratuitous runtime errors at the end of FLASH runs.

This is change set r25841 from the "FLASH3" subversion repository of flash.uchicago.edu
applied as a patch to the FLASH-subset git repository.