Check whether localNumBlocks > MAXBLOCKS to avoid buffer overruns
In the shorter term, the use of MAXBLOCKS in the limited_hdf5_io
impementation files can be eliminated easily, e.g. by using allocatable
buffers. I just don't want to deal with it at this moment.

In the longer term, we should probably get rid of MAXBLOCKS
completely, at least for Mode 3+, since in does not make much
sense to have a maximum number of blocks per proc then.