a first 3D test for Sedov with "unsplit" Hydro
The parfile is essentially the same as for an old test we used
to run regularly in the days of FLASH 3, I think.

See top of parfile for how to configure.
You probably have to increase -maxblocks for running on 1 proc -
doubling it should be more than enough.

This test will create all leaf blocks fully refiend at level 3.
So it is what you @jared321 call PseudoUG.

This test is very low resolution and runs for only 34 time steps.
That's how we used to do things...

This test does not rely on sedov4.5 commits. So it can be run (after
merging of this unsplit3D branch) today in either the development or
the master branch.