get unitTest/Eos working in 3D and 2D with Paramesh
Eos unit test was broken in various ways.

Internally the version of Eos_unitTest that takes 4..5
arguments is called Eos_unitTest4, this is not visible in
the filename physics/Eos/unitTest/Eos_unitTest.F90.
One configuration will only nedd one of the specific
subroutines for which the name Eos_unitTest is now
generic, so this should not cause conflicts.

Note: this will need fixes for index reordering, and testing with
Mode3.  Hopefully to follow soon in another commit.