Add very simple EosGraph setup from FLASH 4
This can be used as a convenient step in creating some 1D curves
showing the behavior of an EOS.

It just writes a couple of text files with some columns of data.
Currently P(T) and E(T), and some other quantities,  at fixed density.

Any adjustments have to be made by editing the included minimal
Driver_evolveFlash.F90.