A variant of hy_advance.F90 that demonstrates OpenMP threading in Mode3
Currently only present as a private overriding version of hy_advance
in the Sedov simulation directory, but this should awork as well for
other setups using the regular unsplit Hydro implementation.  The
simplest way to enable this for testing is to add +threadBL to the
setup command. (This enables compilation with OpenMP [and turns block
list threading on by default] -- nothing new here -- and thus requires
support for the relevant compiler flags in the Makefile.h being used.)

I had to resort to using the gr_iterator_t with gr_getBlkIterator, etc.,
rather than using the leaf_iterator_t with Grid_getLeafIterator, etc.,
because the latter (in its current form based on amrex_octree_module)
did not work properly; I did not dig deep to figure out why.

Based on limited testing on asterix, this code appears to work correctly
in Mode3 in two cases (via different code variants in hy_advance):

  1. flux_correct=.FALSE.                               (FIRST VARIANT)
  2. flux_correct=.TRUE., flux_correct_perLevel=.FALSE. (THIRD VARIANT)

(The remaining case flux_correct=.TRUE. and flux_correct_perLevel=.TRUE.
was not supported in Mode3 anyway.)

Note that several tweaks in previous commits, ancestors of this one,
are required to make this work.)