Do not any more zero-fill flux arrays in various places
If our indexing logic is otherwise right, we should not have
to pre-initialize the data in flux arrays to zero.  The Riemann
solver and other code in hy_getFaceFlux should properly initialize
the fluxes at all location where their values will be used; and
where their values are NOT used (as in guard cell areas that are
part of the arrays that hold the fluxes), there is no need to
initialize to 0.

Also a minor comment tweak.