Fix gr_estimateBlkError for PARAMESH with spherical coordinates
Also applies to some other combinations of geometry and dimensionality
in the PARAMESH case (but not Cartesian or 2D cylindrical geometry).

This avoids using gr_oneBlock(blkID)%firstAxisCoords and
gr_oneBlock(blkID)%secondAxisCoords arrays, which are not very useful
in a world where the convention for cell coordinateds is a global one.

The handling of coordinate information is now the same for PARAMESH
as with AMReX is this routine.

Some allocations and Grid_getCellCoords calls are now skipped when
unnecessary because of simple geometry.