Revert "Unnest the auxiliary function flatten()"
This reverts commit 5c3567327aec97156640051073454ead9263e7ec.

Apparently that "fix" did not work. It has the result that Config
files that use the ##python:genLines mechanism are not properly
processed. I need to understand better how the genLines machanism is
supposed to work, before having another go at a "fix" that was
(apparently) only needed for some versions of Python2 anyway!

Signed-off-by: Klaus Weide <klaus@flash.uchicago.edu>