(Issue #96) Add early compile-time test whether N_DIM==AMREX_SPACEDIM
Produce an error if the proprocessor finds N_DIM != AMREX_SPACEDIM .

This assumes that N_DIM and NDIM are defined to the same value and that
the definition of N_DIM is passed on the compiler command line, as
is normally ensured by the FLASH build system.

This of course also assumes that the compiler is invoked with the
correct command line (in partucular, -I flags) so that the header
file AMReX_Config.H is found.

This check could have been added to many files instead of (or in
addition to) this one, gr_amrexInterface.F90. I chose this one
because (1) it is automatically compiled if the AMReX library
is used to provided the GridMain implementation (but not otherwise),
and (2) this file should usually be compiled fairly early by
GNU make.

(cherry picked from commit f1f2c1f46b4ee9b15561f0a5137390e1fac2a1f9)