Minimal DustCollapse changes, some doc updates
These changes in the DustCollapse simulation are fairly obvious,
something like this is needed to update the simulation to the
new conventions.

These changes alone are not sufficient to make a working
setup for Amrex AMR (but may be sufficient with paramesh).
At least the gr_markRefineDerefineCallback.F90 will have
to be updated or replaced.

Also minor doc additions in Grid_markRefineDerefine instances.