Change io_readData to include "Flash_mpi_implicitNone.fh"
This source file should now be ready to work in all three scenarios:

1. if HAVE_MPIF08_MODULE is defined (and is not lying);
2. if HAVE_MPIF08_MODULE is undefined and HAVE_MPI_MODULE is defined (and is not lying);
3. if neither is defined (but MPI Fortran bindings work using mpif.h ).

This file has been adapted as an exercise and example.
One may not want to convert all existing files that use MPI
calls in a simular way. One may not even want to convert
all such files in a similar way for which this is easily
possible (e.g., no nonblocking calls or request objects).
But one may at least decide to write _new_ code that has
MPI calls following a similar pattern, to get the benefits
of mpi_f00 especially for debugging (type checking).